requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
chromadb
transformers
sentence-transformers
pypdf
python-dotenv
huggingface_hub
streamlit
